reinforced user ownership validation in basic edit/remove operations

+++ api/public/controllers.autoload/Accounts.php
@@ -155,6 +155,13 @@ class Accounts
 
             $userID = UserModel::getUserIdByName($authusername, true);
 
+            // Make sure account belongs to user
+            if (!AccountModel::exists([
+                "users_user_id" => $userID,
+            ])) {
+                throw new BadValidationTypeException("Account not found!");
+            }
+
             TransactionModel::delete([
                 "accounts_account_from_id" => $accountID
             ]);
